IMoralUtil: Removing and renaming moral to morale
